refacto(air/layout): move compute_program_hash to layout/mod.rs
