+++ crates/air/src/layout/dex/mod.rs
@@ -4,8 +4,7 @@ pub mod global_values;
 use crate::{
     consts::*,
     felt_hex,
-    layout::stark_curve,
-    layout::utils::compute_program_hash,
+    layout::{compute_program_hash, stark_curve},
     periodic_columns::{eval_ecdsa_x, eval_ecdsa_y, eval_pedersen_x, eval_pedersen_y},
     public_memory::{PublicInput, INITIAL_PC, MAX_ADDRESS, MAX_LOG_N_STEPS, MAX_RANGE_CHECK},
 };
