+++ crates/air/src/layout/mod.rs
@@ -1,11 +1,14 @@
-use crate::{domains::StarkDomains, public_memory::PublicInput};
+use crate::{
+    consts::{FELT_0, FELT_3},
+    domains::StarkDomains,
+    public_memory::PublicInput,
+};
+use alloc::vec::Vec;
 use num_bigint::{BigInt, TryFromBigIntError};
 use starknet_core::types::NonZeroFelt;
-use starknet_crypto::Felt;
+use starknet_crypto::{pedersen_hash, Felt};
 use starknet_types_core::felt::FeltIsZeroError;
 use swiftness_transcript::transcript::Transcript;
-
-pub mod utils;
 
 #[cfg(feature = "dex")]
 pub mod dex;
@@ -110,6 +113,24 @@ pub fn safe_mult(value: Felt, multiplier: Felt) -> Result<Felt, SafeMultError> {
         core::cmp::Ordering::Equal => Ok(felt_mul),
         _ => Err(SafeMultError::Overflow { actual: felt_mul.to_bigint(), expected: mul }),
     }
+}
+
+pub fn compute_program_hash(
+    memory: &[Felt],
+    initial_pc: Felt,
+    initial_fp: Felt,
+) -> Result<Felt, TryFromBigIntError<BigInt>> {
+    let program: Vec<&Felt> = memory
+        .iter()
+        .skip(initial_pc.to_bigint().try_into()?)
+        .step_by(2)
+        .take((initial_fp - FELT_3).to_bigint().try_into()?)
+        .collect();
+
+    let hash = program.iter().fold(FELT_0, |acc, &e| pedersen_hash(&acc, e));
+    let program_hash = pedersen_hash(&hash, &Felt::from(program.len()));
+
+    Ok(program_hash)
 }
 
 #[cfg(feature = "std")]
